chroe: move examples traits in common place
