+++ examples/CMakeLists.txt
@@ -12,6 +12,10 @@ foreach(example ${examples})
     set(example_target "${examples_global_target}.${example}")
 
     add_executable(${example_target})
-    target_sources(${example_target} PRIVATE ${example}.cpp)
+    target_sources(${example_target} PRIVATE ${example}.cpp common/traits.hpp)
+    target_include_directories(
+        ${example_target}
+        PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}
+    )
     target_link_libraries(${example_target} PRIVATE beman::monadics)
 endforeach()
